Update pagination&form input and change api. Create bookmark, readinglist, promt novel method

+++ novel-service/src/main/java/com/backend/novelservice/dto/response/NovelResponse.java
@@ -28,7 +28,7 @@ public class NovelResponse {
     String authorId;
     String authorName;
     String description;
-    Integer score;
+    double score;
     Integer bookStatus;
     Long visitCount;
     Long wordCount;
